pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "amanogawa"
version = "0.1.0"
description = "Quantify Milky Way stellar clustering & dark lane morphology from a single image"
readme = "README.md"
requires-python = ">=3.10,<3.13"
license = { file = "LICENSE" }
authors = [
  { name = "Shunya Hayashida" },
]
dependencies = [
  "numpy==1.26.4",
  "scipy==1.13.1",
  "scikit-image==0.24.0",
  "pandas==2.2.2",
  "matplotlib==3.9.0",
  "astropy==6.1.0",
  "scikit-learn==1.5.1",
  "photutils>=1.11.0",
  "pyyaml==6.0.1",
  "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8",
  "pytest-cov>=5",
  "ruff>=0.5",
]
heif = [
  "pillow-heif>=0.18.0",
]

[project.urls]
Homepage = "https://github.com/SHayashida/Amanogawa"
Repository = "https://github.com/SHayashida/Amanogawa"

[project.scripts]
amanogawa-detect = "amanogawa.detection:main"
amanogawa-stats = "amanogawa.spatial_stats:main"
amanogawa-band = "amanogawa.band_geometry:main"
amanogawa-dark = "amanogawa.dark_morphology:main"
amanogawa-photometry = "amanogawa.photometry:main"
amanogawa-run = "amanogawa.run:main"
amanogawa-fits-export = "amanogawa.fits_export:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["amanogawa*"]

[tool.ruff]
line-length = 100
target-version = "py310"
extend-exclude = [
  "data",
  "notebooks",
  "outputs",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
